backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
paho-mqtt>=2.1.0
scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=2.0.0
joblib>=1.3.0
